+++ packages/atomic/src/components/atomic-no-results/atomic-no-results.stories.tsx
@@ -1,11 +1,9 @@
 import defaultStory from '../../../.storybook/default-story';
-import NoResultsDoc from './atomic-no-results.mdx';
 
 const {defaultModuleExport, exportedStory} = defaultStory(
   'Atomic/NoResult',
   'atomic-no-results',
   {},
-  NoResultsDoc,
   {
     engineConfig: {
       search: {