+++ packages/atomic/src/components/facets/atomic-numeric-facet/atomic-numeric-facet.stories.tsx
@@ -1,12 +1,10 @@
 import {html} from 'lit-html';
 import defaultStory from '../../../../.storybook/default-story';
-import NumericFacetDoc from './atomic-numeric-facet.mdx';
 
 const {defaultModuleExport, exportedStory} = defaultStory(
   'Atomic/NumericFacet',
   'atomic-numeric-facet',
   {field: 'ytviewcount'},
-  NumericFacetDoc,
   {
     additionalMarkup: () => html`<style>
       atomic-numeric-facet {