+++ packages/atomic/src/index.ts
@@ -1,3 +1,14 @@
 export * from './components';
 
 export * from './utils/result-utils';
+
+export {
+  initializeBindings,
+  Bindings,
+  MissingInterfaceParentError,
+  BindStateToController,
+  InitializeBindings,
+  InitializableComponent,
+} from './utils/initialization-utils';
+
+export {ResultContext} from './components/result-template-components/result-template-decorators';